chore(ai): extract layout/template prompts [refs #363]
